Set base type of user-defined types to BYTEAARRAY
`CREATE TABLE` on columns of user-defined type has issues with trying to
figure out the base type.  Consider all such failures of finding the
base type to be from user-defined types, so then interpret the base type
as `BYTEAARRAY`.  Choose this base type because it translates to DocDB's
`YB_YQL_DATA_TYPE_BINARY` and does not `allow_for_primary_key`, both of
which are desired qualities.