Support functions returning record type
The `yb_plpgsql` test has a part where a function declares a variable of
record type and returns it as a UDT.  This part crashes the server
because it encounters a new code path.  In that code path, a temporary
heap tuple is not fully invalidated because it doesn't invalidate the
`t_ybctid` field.  Explicitly invalidate that field of the heap tuple.
Do the same for another part of the code in the same file.  The server
should no longer crash in the `yb_plpgsql` test.