Add base type test and handle more base type cases
Create a test for user-defined base types because there are issues with
the implementation of base types.  Use the test to catch cases, and fix
those cases.  The test (and fix) should cover variations of `typlen` and
`typbyval` parameters, with an additional, but less-emphasized,
variation on `typalign`.  Also update `yb_pg_types_udt_serial_schedule`
to be organized by categories of user-defined types.