Improve base type test with custom C func type
Define custom input and output functions for defining a new `bigname`
type that is of fixed length 128 bytes.  To get the tests to pass,
update `YBCFixedLengthByRefTypeEntity` to be of size 128 bytes as well
(this is not the intended end solution).  Also test the `json` type and
add some more selects for the user-defined `name` type.  The
`TestPgRegressTypesUDT` java test should still pass.