Remove tablet from TableInfo of system tables
When deleting a database, all tables go through
`DeleteTabletsAndSendRequests`.  This poses a problem for system tables
because we do not want to delete the system catalog tablet.

Remove the in-memory reference to the system catalog tablet in the
`TableInfo` for system tables _before_ calling
`DeleteTabletsAndSendRequests`.  This should prevent the deletion of the
system catalog tablet.