Fix marking system tables for deletion
The previous commit changes the way system tables in
`DeleteYsqlDBTables` are handled by not removing the system catalog
tablet from the in-memory `TableInfo` but at the same time not deleting
the system catalog tablet.  This is not fine when we get to
`CleanUpDeletedTables` because it checks for `AreAllTabletsDeleted`.
Relax the condition to also allow system tables to go through.  This
should be fine because a previous condition makes sure that tables are
in the `DELETING` state.