Fix not running BG task for delete empty DB
When deleting an empty database, no tablets are deleted.  Then,
`CatalogManagerBgTasks::Run` does not find tablets to delete, so it
doesn't run `CatalogManager::CleanUpDeletedTables`.  This leaves the
system tables of that empty database in the `DELETING` state.  Only when
a later deletion actually deletes a tablet,
`CatalogManager::CleanUpDeletedTables` gets triggered, and the tables of
the empty database finally transition to the `DELETED` state.

Fix this by adding a function `CatalogManager::AreTablesDeleting` to
check whether any tables are in the `DELETING` (or `DELETED`) state.
Use this in disjunction with the existing
`CatalogManager::ExtractTabletsToProcess` check to determine whether to
run `CatalogManager::CleanUpDeletedTables`.

To prevent the the `CatalogManager::AreTablesDeleting` check from
continually returning true due to deleted tables not removed from
`table_ids_map_`, actually remove tables from the map (commented out by
@bmatican).